Move reusable functions like hidePage, showPage, and purgeKarma to their own modules in utils/
