+++ src/constants/karmaIdentifiers.js
@@ -1,0 +1,9 @@
+const karmaIdentifiers = {
+  hideKarma: ['.score', '.karma'],
+  hideAwards: ['.awarding-link', '.awarding-show-more-link'],
+  hideVotingButtons: ['.arrow'],
+  hideUsernames: ['.author'],
+  hideOwnKarma: ['.userkarma'],
+};
+
+export default karmaIdentifiers;
