+++ src/utils/showPage.js
@@ -1,0 +1,3 @@
+export default function showPage() {
+  document.documentElement.style.filter = 'none';
+}
